add a Makefile; move unused files out of the way
I'm tired of waiting for everything to be recompiled after every little
change. It's much faster to generate an object file for each source
file, update each of those only when its source file changes, and relink
each time.

Linking libft8.a explicitly is still ugly, but works for now.