Strip <> brackets from callsign before replying
The brackets just mean the callsign is hashed. If I see a message like
  JH6URJ <CQ35MD> -13
and click on it, I expect to attempt to call CQ35MD (although it's a
bit premature: better to wait for him to finish his QSO). Keeping the
brackets otherwise results in an error "Cannot parse message!"

Another time I clicked on the callsign <OK1PHM/QRP> and got this error:
  *Error: field[#contact_callsign] can't be set to [<OK1PHM/QRP>], improper size.
The field is 11 characters max, and that callsign now fits without the
brackets (as it should: up to 11 characters are allowed for FT8 callsigns.)