Log tx power and swr; ADIF export; show power in 1/10 units
There's no ADIF field defined for SWR, AFAICT.

I changed the data/create_db.sql script, but of course I didn't want to
start over with a fresh database; so I did

  $ sqlite3 ~/sbitx/data/sbitx.db
  sqlite> alter table logbook add column tx_power text default "";
  sqlite> alter table logbook add column vswr text default "";

That puts the columns in the "wrong" order: they come last, but I'd
rather have the comment field last.  Verify it like this:

  sqlite> .schema logbook

But the code should not depend on column order: i.e. it should not do a
query like 'select * from logbook ...' and then expect the columns to be
in a particular order, for display purposes or otherwise.

This code could all be improved quite a bit, but export_adif() is
particularly bad: it iterates fields and expects column 1 to be mode,
7 to be "my gridsquare" and so on. And hard-coded numbers too: nobody
even bothered with an enum for that! So to keep that working, the query
now needs to specify the columns to fetch. But it could alternatively
loop over columns and use the column name to decide which ADIF field to
use: that might be more robust (and also somewhat less efficient).
I didn't test the rest of the logbook functionality yet, only did an FT8
QSO and exported ADIF; so other things could be broken.