Don't call a path a "directory" when reading settings
It implies a very basic misunderstanding. And it's wasteful to keep this
path string around longer than we need it, so put it in a block.

If one searches for "getenv", probably most of those places could
have better implementation of constructing path strings, but most are
not as egregious as this one.