draw_text: remove support for style markup; return length in pixels
Amends 7c49b7649b6bb330c6b2ee749c079bccff938628 (decrufting)

The length is useful to avoid having to call measure_text() separately,
which ought to reduce the total number of cairo_x() function calls
in the case that it's needed (and otherwise may be slightly wasteful).
But not drawing characters one-by-one ought to be a bigger win.