FT8: add separate style for "callee" callsign
Prior to my revamp, it would have been in STYLE_FT8_RX style. Then it
started highlighting both callsigns with STYLE_CALLER: just wrong.

But so far I did not teach ft8lib to distinguish caller and callee;
so we have to count them out from the total number of callsigns found.
Perhaps libft8 should distinguish them: that will be handy if it
supports multi-part messages later on.

Anyway, IMO it's nice to show the callee in a dimmer color if it's
not my callsign, because the UI does not make it easy to call that ham
anyway (you'd have to type it in by hand, you can't click on it).