9p: use pthread; can read the frequency!
One small step for a service, one giant leap for mumble mumble

Forking was a bad idea, because that just kicks the can down the road:
we want 9p as an IPC mechanism, but we'd need another IPC mechanism to
communicate with the main sbitx process. This way memory is shared,
and it's ok to call get_field_value() to get current values (rather
than the stale ones that the fork would have).