get_console_text: don't count lines that don't match filter; tail == cat
The lower loop was skipping lines that don't match the filter, as it
should; but the upper loop was using their start_column and len directly
to decide which line to start on. After each successful read_text,
it gets called again to check for leftovers, and it must not return
any more text if there's no more to read; in fact it was usually
duplicating some.

Now the result of `tail -f` matches the result of `cat` and the
lines found in display_log.txt: i.e. `tail -f` is finally useful.