FT8: Popover to initiate a call, after a long-press or encoder press
Until now, it has been too easy to accidentally start calling an
unintended callsign, due to fat-fingering the touchscreen and getting
the wrong line, or touching while the console is scrolling up. Now, if
you hold in one position for 800ms or more, it copies the callsign found
to the CALL field, opens a popover with a "Call" button on it, and you
have to tap that button to initiate the QSO.

Another oddity before was that you could use the upper encoder to scroll
through the console, but pressing the encoder's button changed focus
back to the audio field: so there was not much use in scrolling that
way. Now, if the console field has focus, a short press on the encoder's
button is the same as a long-press on the console line: the same popover
opens, so that you can confirm whether you want to start a QSO.

On the downside: this slows down initiation a bit, and might make
continuing a QSO from a non-initial state somewhat harder. But we still
have the buttons across the bottom for that.

Another downside: popovers are not easy enough to dismiss in GTK 3.
You can tap on the margin area (so the top margin is larger, to make
that easier).  GTK4 has the autohide property, so you could close by
tapping outside; but GTK3 doesn't.  We also compensate by having
the popover disappear the next time a line is added to the console.
The popover is meant to "point to" the relevant callsign on the
relevant line: so if the console scrolls up, it would no longer be
pointing to the right place. Automatically dismissing it after
15 seconds also avoids the chance that someone will complain that they
don't know how to get rid of this popover after it pops up.

The console is not scrollable, but now it could be made scrollable:
a flicking gesture would not necessarily be confused with a tap.
But this console is not even an independent widget, so it may not
be so easy to use whatever GTK has for that.