9p: Hold console-last-line constant after spans opened, per client
A client is expected to open the spans file (metadata) first, if at all.
In fs_open(spansfile), memorize the "console last line" (aka data_index,
as a general concept) by storing it in the FidAux. Next, when the same
client calls fs_open(textfile), look up the "console last line"
(data_index) in the FidAux corresponding to the spans file that the
same client has already opened.

To make that possible, we need a client ID. It turns out that if we set
the void* Ixp9Req::Ixp9Srv::aux (r->srv->aux), it will come back in each
subsequent request. For now, rather than an actual pointer, just set a
numeric client ID (0xa44a............), "ATTA0..0xx", in the order that
they attached. Then in the second fs_open, we can look up the FidAux
(findfidaux(r->srv->aux, start_from)) in which the file ID (QID,
FidAux::Devfile::id) matches the pattern for channel text spans.

In order to make _that_ possible, it's better to add some QID enums
to make the numeric pattern explicit. (But the pattern may change:
we will have more kinds of text/span combinations, i.e. in different
modes such as CW, and also for the whole-console file; and it needs
to be easy to identify that they go together in all those cases.)

In order to make _that_ reasonable, it's better to rearrange the file
to keep the structs, enums and devfiles table close together.

So this turned into a big refactoring.

It's hard to prove that it works, so far, because the console is not
changing very often in FT8 mode (only every 15 seconds), so the odds of
a collision (client such as mtail reading while write_console_semantic()
is busy changing the console text) are quite unlikely in the first place.
But the debug output looks correct.

There is no mutex, just an assumption that if the client is reading
back from the "bottom", and we fix what "bottom" means temporarily,
the client is unlikely to read far enough back to run into changing
data, because the console is ring-buffer-like: records do not move,
they are only overwritten. (But so far it's not quite a proper ring
buffer: I think that after a console wrap-around, the client will read
back to the top, and not be able to go any further back: so it will not
be able to read as many lines as usual, and in fact some of the newest
batch of FT8 messages may be missed. This needs fixing separately.)

So the "unlocking" needs to occur by setting data_index back to -1 when
the server detects that the client is done with the critical part of
reading, in some sense. mtail first reads all the spans that it's
interested in (i.e. if it wants 10 lines of history, it will read spans
until start_row == tail start_row - 10), before it begins to read text
lines; so if we expect that access pattern, it's ok to "unlock" after
stat_text(). The client uses stat() to find out how big the text is, and
we might tell a lie: we might say it's shorter than reality, to be
consistent with the result of stat_text_spans() that we have previously
returned to the client. After that, the client is expected to pread()
(or fseek() and then read()) with a suitable offset, going backwards
from what it sees as the "end". But what if the client calls stat()
multiple times, for some reason?  So to be safe, maybe it's better not
to unlock until the read("received") begins. That's what we do for now.

Debug output has "qid 0x..." alongside the name in some cases, now
that they are more meaningful in hex than in decimal.