9p: fix bugs for mtail received; increment row number forever
The 9p client does not know or need to care how many console lines are
stored: mtail in particular just expects to read the changed lines each
time the file size changes, which looks the same as if the file kept
growing, as if all history was stored indefinitely. So the row number
(as reported in each span) should increment indefinitely too.