Make it build with plain "make"
What I ran into first when I tried building my fork on zbitx:

I guess I was getting away with some problems by always building the
debug version on sbitx. But in general, if a case in a switch
declares local variables, it must be a block (surrounded with braces).

Don't require libixp in main branch. (I haven't got it installed there
yet; but it will soon be needed anyway on the 9p branch.)