Fix local variables in case statements
What I ran into first when I tried building my fork on zbitx:

I guess I was getting away with some problems by always building the
debug version on sbitx. But in general, if a case in a switch
declares local variables, it must be a block (surrounded with braces).