Start work with zbitx
Based on Ashar Farhan's patch 'zbitx i2c interface'
686ac568e820ca20ff6f94c4c95e34756590f146
which actually does many zbitx things, not just the i2c interface.
Merged into K7IHZ fork with a good bit of adjustment.

In particular:

void rx_linear() has
```
	// gather the samples into a time domain array
	for (i = MAX_BINS / 2; i < MAX_BINS; i++)
	{
		i_sample = (1.0 * input_rx[j]) / 200000000.0;
		q_sample = 0;
```

In Farhan's patch it's changed to 20000000.0 and that makes the radio
much more sensitive, perhaps too much so. Setting IF gain to 50 looks
like too much already, with that change. If this change should really
be there, it should be in another patch, explained well.