Get rid of zbitx_available, use sbitx_hw_version enum; fix power and swr overwriting
read_power() is for sBitx, not zBitx. But if the "hw" setting were in
hw_settings.ini, tx_process() would not call read_power(). But it was
missing from my hw_settings.ini; and anyway, I think it's best to detect
hardware rather than relying on an .ini file. And in fact, so far, if
the Pi Pico touchscreen is detected, we've got a zBitx, so just override
the sbitx_hw_version accordingly.

On the zBitx, cmd_exec() calls set_field("#fwdpower") when the Pico pushes a
power reading.

query_swr() is dead code AFAICT, so remove it.