Hold last-reported power and swr for the logbook
On the zBitx, the Pi Pico firmware sends power and swr updates when
polled, and we store them in the #fwdpower and #vswr fields right away.
So they are usually 0 at the time logbook_add() is called; so we need a
memory of the previous non-zero values.

Amends 81efc24c25ccf28c871f3cbc8cd4c879bd0bdda7