9p: Support FT8 transmit
% echo 'N1ABC K1DEF AB12' > /n/sbitx/modes/ft8/1/send
schedules transmission, starts it right way if it's in the timeslot.
% echo '' > /n/sbitx/modes/ft8/1/send
stops transmission right away, and also clears the #text_in field.

Hitting ESC on the GTK UI always seemed a bit too drastic, forgetting
the contact info completely. This way though, the remote UI has a
choice: stop transmitting but retain the message text locally in case
the user wants to try again, or stop and clear everything. Clearing
the other QSO fields has to be done separately anyway.