Optimize waterfall data somewhat
It might be a good idea to cache more than one line of waterfall data,
but in fact the code is not doing that. In fact, the GTK UI annoyingly
loses waterfall history at the drop of a hat: e.g. even if you only move
the window and don't resize it. And after converting the raw data to
false color, the history is also stored in that form. So it doesn't
seem useful to malloc the `wf` buffer, nor for it to be bigger than
one line's worth. It's now a static array of 1024 ints and I seem
to be getting away with it, even with ASAN turned on.

Also why not cache the bool from the AUTOSCOPE field rather than
redoing the necessary string comparisons during calculation of
EACH INDIVIDUAL WATERFALL DATA POINT...