9p: when frequency is set, set the mode too
...and for that, write_field() needs the whole Devfile struct, not only
the field name, just as read_field() needs it.

This is a workaround for when I want to switch to FT8 from the remote UI
and it happens to be in a different mode. Hopefully some day it will
be able to do all the modes at the same time: but there might still be
some use for telling the radio that the user is focusing on a particular
channel.