events file gives paths and values; more complete
tab-separated subpath and value, e.g.
  battery/voltage	727
  s	1001
  text
  modes/ft8/1/received

i.e. if the event line contains a tab character, then the client can
avoid needing to open and read the value file: it's already given.
But we don't do that for longer content like text and received.

notify_field_changed() is called from more places than I'd like, but
removing any of them breaks something. events are deduplicated anyway.

Devfile::read_name has to be unique now: e.g. we use r1:freq-max for
modes/ft8/1/frequency.meta/max to avoid sending that as the event path
when the value of modes/ft8/1/frequency changes. Most of these
metavalues are constant anyway. modes/ft8/1/frequency.meta/step isn't
constant but it also has its own command/id. But the convention is to
use hyphen as the separator, in case we need to tokenize it later for
some reason.  Some IDs already have underscores in them, and this
separator should be unique.