Move s-meter and drive into channel; add metadata
The S-meter tells strength of one signal, so if we are receiving multiple
simultaneous channels, each would have a different signal strength.

As with band-specific IF gain, we should have band-specific drive.
If the UI has channels open on multiple bands, it won't be able to
work them all at the same time, but focusing one channel ought to switch,
and then it becomes relevant that the drive is a channel setting.
Anyway, maybe one wants to use less power on FT8 than on Olivia,
or something like that.