Popover call button: just call the extracted callsign; remove dead code
If I long-press a line with something like
CQ POTA A1XYZ
and hit the popover call button, ft8_process(msg, FT8_START_QSO) sends
POTA LB2JK JO59
whereas if the callsign is shorter, it works. But the point of extracting
the callsign ahead of time to show in the popover and in the CALL field
is to be sure who we're going to call when hitting the Call button; so
just compose a message right there, as the F2 macro button would do
(although in this version it's hard-coded, not using the macro).

Remove get_ft8_callsign(): it was added in
4a98fe3b0b97ebcfcc726e7cb9d6c29c82727293 and apparently never used.