Ensure that we never send an empty R report
It seems the usual sequence is supposed to be
ft8_message_tokenize():
	strcpy(signal_strength, p);
ft8_on_start_qso():
    field_set("SENT", signal_strength);
ft8_process():
	report_send = field_str("SENT");
ft8_on_signal_report():
	sprintf(reply_message, "%s %s R%s", call, mycall, report_send);

But what if ft8_on_start_qso() didn't get called? I've seen cases where
I send out a useless packet like "A1XYZ LB2JK R"
If sbitx receives such a packet, it treats it as R+00; but maybe most
software will see it as invalid. I doubt such QSOs get logged.
So make sure that if "SENT" is not set, we set it to the signal strength
of the last packet (so that it's visible in the GTK SENT field before
we actually start sending the packet).