Fix FT8 state machine to deal with hashed callsigns; store sent callsign hashes
The hashtable in my sbitx doesn't even contain the hash of my own callsign
unless I let ftx_message_encode() save it (as if it wasn't already clear
that when an ftx_ function takes hash_if, it should be provided, not NULL!)
And if I cold-call a callsign, I need to store his hash to be ready to
handle his replies, too. This becomes quite noticeable when either
callsign is nonstandard and needs to be sent hashed.

So far we still show <> brackets around hashed callsigns in the console,
just to be able to see that it's hashed; but for running the state machine
the <> brackets need to be removed; so add a tokncpy() function to copy
from messages into the m1..m4 variables (which then also has to deal with
"tokens" like R+12, R-20, and callsigns with slashes).