Popover: extract more fields from selected message; use new ft8_call()
When doing the long-press or right-click, extract as many fields as
possible: time, callsigns, grid, RST, and check for matching mycallsign.
Hide the Reply button unless the message contains my callsign.
When hitting the Call button, do not re-parse the message: rather
call the new ft8_call() function, which relies on relevant fields
having been set (the CALL field is the most important). Set the
TEXT field, in case the user wants to edit it (although then there
would be the other annoyance that hitting ESC erases everything,
and we don't have a separate button to just stop transmitting but
retain the fields).

This makes it look a bit more transparent / deterministic: the fields
are populated before the user hits Call, so it's fairly clear what
sort of message would be sent out.