ft8_tx_3f: don't copy the outgoing message to the TEXT field
In other places in the code, those fields are treated as "the logger",
and 4a98fe3b0b97ebcfcc726e7cb9d6c29c82727293 changed enter_qso()
to log the TEXT field as a comment. So I was getting the final message
logged as a comment every time: either an RRR or a 73. If the idea
is to type something there as a comment, overwriting it is not good.

In fact, this field has a secondary purpose though: you can type a
full FT8 message and hit enter to send it. In
206df29b79fc5ee4ddf088af2debab89570ce6b5 I wanted to make it possible to
abort an outgoing message, edit it, and send again, in case the state
machine gets it wrong. I guess we need a "copy" function for that
purpose, then.