Make popover available on non-incoming lines, with Copy function
In case we want to abort an outgoing message, edit and send again:
hit esc, long-press or right-click on the line to copy it to the
TEXT field, then edit it and hit enter when done.

Reusing one popover for too many purposes is getting messy. But I
suppose it saves a little memory to avoid creating more than one.