Right-click / long-press sets selection; popover Copy copies to clipboard
In _any_ mode, right-clicking or long-press on a console line sets the
X11 selection, so that you could then middle-button-click to paste that
whole line somewhere else. E.g. that might be useful in CW mode...

If you get a Copy button on the popover, it copies the text both to the
TEXT field and to the clipboard, so you could then control-V to paste it
into another application.