Compensate zbitx receive sensitivity: different divisor in rx_linear()
As described in 526b452459df527e2a4e8f9a393e999f3886307e,
Farhan's patch 686ac568e820ca20ff6f94c4c95e34756590f146
changed the divisor without explanation. But with his code,
setting IF gain somewhere under 40 was practical; whereas if we
keep the same divisor as on sbitx, one needs to turn it up much higher,
around 90 or so. It's best to avoid running out of range, in case
one really needs more sensitivity that usual, in some situation.

Now I compared the appearance of the waterfall on sbitx vs. zbitx
with the same antenna, swapping back and forth between radios, while
having both tuned to 14.074 and both having the same waterfall settings
on menu 2: scopegain 1, scopeavg 2, intensity 8, autoscope off.
It seems that IF gain around 60 is sortof reasonable under these
conditions, on both radios, if the divisor on zbitx is half the
value of the divisor on sbitx. So basically zbitx hardware is half
as sensitive on a voltage scale, and we compensate for that by
dividing by a number half as big.

It would be better to calibrate both radios more accurately somehow.
S meter readings should mean something, but in fact I see S3 with no
antenna connected, on both radios, perhaps just from the birdies.