Don't record power if it seems to be tapering off at the end
On the zbitx, the Pi Pico measures outgoing power level and sends us
updates periodically. Often it happens that the final power reading
occurs at the end of the transmission, when it's already tapering off.
That is not what we want to record in the log, so just ignore that one.