Tapping console line sets X11 selection and calls new ft8_call()
Extract the callsign to call, and the other available data (grid, signal
strength, time) from the console semantics, rather than re-parsing the
whole line. I.e. we know which callsign we are going to call because
ft8lib told us which range of characters is the "de" call, not because
we just picked the second word in the message and hoped for the best.
Messages like "CQ POTA X1ABC AA01" will not throw it off: ft8lib knows
the callsign is X1ABC, not POTA.

ft8_call() depends on fields having been set, but that makes behavior
a bit more predictable: the fields can be checked before calling.
Not that the code is making any use of that yet...

It bothers me that it's too easy to tap the wrong line on the
touchscreen and start making an unintentional call, or even call myself
accidentally. So on other branches, there are patches to add a popover
window offering more information and options: last recorded QSO with
this callsign, and buttons to reply to a message from a QSO in progress,
start a QSO, or just copy the text to the clipboard. However, GTK 2
popovers seem buggy: you can't dismiss it by clicking outside, so I
tried to dismiss it automatically when the console scrolls; but
sometimes that causes a crash in the GTK library:

[xcb] Unknown sequence number while processing queue
[xcb] Most likely this is a multi-threaded client and XInitThreads has not been called
[xcb] Aborting, sorry about that.
sbitx: ../../src/xcb_io.c:269: poll_for_event: Assertion `!xcb_xlib_threads_sequence_lost' failed.

Thread 1 "sbitx" received signal SIGABRT, Aborted.
__GI_raise (sig=sig@entry=6) at ../sysdeps/unix/sysv/linux/raise.c:50
50      ../sysdeps/unix/sysv/linux/raise.c: No such file or directory.
(gdb) bt
    (fmt=0x7ff6aa1968 "%s%s%s:%u: %s%sAssertion `%s' failed.\n%n", assertion=assertion@entry=0x7ff6834e80 "!xcb_xlib_threads_sequence_lost", file=file@entry=0x7ff6834d40 "../../src/xcb_io.c", line=line@entry=269, function=function@entry=0x7ff68351a8 "poll_for_event") at assert.c:92
    (assertion=0x7ff6834e80 "!xcb_xlib_threads_sequence_lost", file=0x7ff6834d40 "../../src/xcb_io.c", line=269, function=0x7ff68351a8 "poll_for_event")
    at assert.c:101

So in _this_ version of the patch we stick with the simple tap-to-call
functionality from before. And I rule out calling myself now.