web: Require an FT8 message to contain 1 or more words, not 3 or more
When receiving e.g. a CQ from a complex callsign, there's not enough
bits for anything else: he can't send his grid. So it was not
appropriate for hd_message_parse() to treat a message with only 2 words
as an error.

write_console_semantic() calls hd_decorate() to prepare the message
to send to the web connection, and hd_decorate() leaves the message
empty if hd_message_parse() returns an error. So a message like
"CQ AB/C1DEF" would not be sent to the web connection. Now it is.