Stop doing NTP queries and RTC updates
This is an operating system concern, not an application concern.
For example, chrony professes to do a better job on systems that have
intermittent Internet connectivity: it learns and tracks the drift better.
But the usual Raspberry Pi OS has regular ntpd already running, and that
is also often good enough. And it's possible to have either one of them
update the RTC as well, since it has a kernel driver.