save_user_settings: copy current freq to the currently selected vfo
27006ea31759ffc0435b587ee4e7bc5c3306eac1 looked like an incomplete
thought, unless get_field() was supposed to have side effects (ick).
So now we actually copy it, to ensure that the radio starts up on the
same frequency as when it was shut down.

Also, don't try to save in ui_tick(): that seems excessive. Might as
well try to minimize rewriting the SSD in case power is lost suddenly.
(Not that there aren't enough writes going on anyway...)