Add filter args to get_console and console_current_length; use in 9p
Now /modes/ft8/1/received includes only the console lines that start
with a span (hopefully a whole-line span!) of STYLE_FT8_RX; and
/modes/ft8/1/sent has only STYLE_FT8_TX.

If we wanted to show them together/interleaved, it would be more hassle
now: maybe I'd want to turn some grouped semantics into bit-flags.
But /text is still everything (the whole console). And later on, if the
radio begins to support multiple simultaneous channels, each console
line will need to have a channel number, and then that would need to
be another way to filter; so we could have /modes/ft8/1/text being
everything on channel 1.