Update mtime on all dirs containing (dirs with) console text
Each time we stat() any directory that contains (even indirectly)
a file with console text (filtered or unfiltered), update its mtime.

The idea was to get `tail -f modes/ft8/1/text` working, but there
may be a bug in the 9pfs client or something. It doesn't seem to
notice this file changing, even though the mtime of root and all
intermediate directories now changes.