Add /text and /modes/ft8/1/text
Both of these are the entire console, for now. Next, I need to look for lines
with FT8 semantics in them; then both stat() and read() will work harder.

`tail -f` seems problematic on Linux with 9pfs: probably the kernel is
caching the stat() results, and I don't know of a way the 9p server can
send a notification that the file has changed. But on Plan 9, it seems that
`tail -f` is calling stat() every few seconds, and results are not cached,
so it actually follows the file, with more delay than I'd really like.

I should apparently report a newer mtime on all ancestor-directories of
a file that has changed. At least the root `text` file changes length
often enough, whereas modes/ft8/1/text does not.