allow more FIDs; clunk and freefid do different things
A FidAux struct is not considered "free" unless it has a nil file ptr.
Usually clunk() is followed by freefid(), but not always; so we should
do unique things in each, to verify that both happened.

It seems that 9pfs on Linux has a bug: it's wasting FIDs, not always
freeing them as soon as I'd think it could. So e.g. I could only change
the frequency 3 times before running out. Having a max 256 should allow
it a good number more times without wasting too much memory on the
server side. (I suppose most server implementations malloc all FIDs, so
it's just a small leak that can go unnoticed.) But either the bug needs
fixing or I should use another client on Linux. Plan 9 as a client is
quite nice about it, by comparison.