web server: handle hotspot-detect.html captive wifi URL
For example if you follow
umatechnology.org/how-to-make-a-raspberry-pi-wpa-encrypted-wireless-hotspot-device/
to run hostapd and dnsmasq to get a wifi access point running on the pi,
and connect to it from iOS, it expects there to be such a portal page
where you can accept T&C somehow. iOS doesn't like that the s/zbitx
is not providing Internet access via its access point (it apparently
tries some Apple servers to check that), so it shows the portal page.
Apparently there is no solution except to remind the user what to do:
https://stackoverflow.com/questions/23553449/remain-connected-to-wifi-network-that-has-a-captive-portal#26058667