9p: reorganize into mixed channels directory
Since the UI will probably present a flat array of (sub)windows, one for
each open channel, it makes more sense for it not to need to explore
channels inside modes recursively.

On the downside, if some hypothetical external daemon wishes to populate
channels for additional modes into the same directory, a union mount
might be needed, and there is a risk of name clash: but the risk is
reduced by the name.type convention (as with most normal files, the
extension gives a hint: 1.ft8 is an FT8 channel, of course. But
renaming it will not change its type.)

It may be that some modes won't support multiple channels though:
so far, one never finds FT8 in use in multiple channels within a
25khz passband, although I suppose it could happen some day.
Then maybe calling it 1.ft8 seems silly, but at least it follows an
extensible convention.