Reduce spectrum_plot to 1024 values; simplify further
Apparently the reason for using only bins 1269 to 1803 (out of 2048) may
be that there is only weak response to real signals outside that range,
and otherwise birdies and noise. But the web UI was trying to grab
a bit more than it should for the 25k setting.

Actually this might need calibration: let's define hz per bin, and
verify experimentally that when receiving 2 signals at different
frequencies, they show up in the right bins; and let's have the
web UI consistent with the 9p view. Not done yet; not clear if it
was done before.