store client_data pointer rather than a fake client ID
Fix old todo: return client_data ptr as srv->aux for this client rather
than just using a recognizable fake pointer

Upsides: this is simpler and more efficient; can reliably detect detach
and clean up
Downside: we have to put client-specific data related to multiple
"files" into a single ClientData struct.  Oh well: not so bad, so far.

One would think on-demand malloc() of ClientData would be good enough,
but the client_data array is actually useful in notify_field_changed():
we need to notify all connected clients, so we'd anyway have to keep
track of the ClientData structs that are allocated. But the rest of the
time, we can get the pointer from FidAux and/or srv->aux, so now we
don't need to search the array as before.