Add xOTA fields: change in settings dialog
This doesn't do anything useful yet, but at least the field values
get saved in user_settings.ini. The intention is to have it export
e.g. my_sota_ref in the ADIF log, and also send an occasional FTx
packet with this location/ref in FT8 and FT4 modes.