Add xOTA fields
This doesn't do anything useful yet, but at least the field values
can be changed from the modified zbitx front panel firmware
https://github.com/ec1oud/zbitxfrontpanel/commit/9b29cf3884ebc4b1eb4df38e928b893818453c75
and get saved in user_settings.ini. The intention is to have it export
e.g. my_sota_ref in the ADIF log, and also send an occasional FTx packet
with this location/ref in FT8 and FT4 modes.