modem_ft8: common time format; try to decide what will be sent; decruft
That is, if the user hits the CQ button and it's not time to start yet,
try to decide whether it will be the regular CQ or the xota alt text,
rather than outputting CQ message to the console and then sending the
xOTA location instead. But with the current approach, it still gets it
wrong sometimes. Maybe we should always look into the future as much
as necessary to decide accurately what shall be sent, rather than
deciding just-in-time in ft8_poll(). And related to that: I don't
think it's useful to start the transmitter if less than half of the
slot time is available; maybe not even if most of the slot time isn't
available. It's a waste of power. So deciding what to do next could be
helpful to avoid that, too.

All time stamps are formatted with tenth-of-a-second resolution now, but
the tenth is omitted if zero (which we often expect in FT8 mode).