Leave notes about FTx commands
This whole thing needs rearchitecting first, to use field IDs as
IDs/commands and use labels only as labels (OTOH we might need to
add a translation solution eventually). E.g. we cannot yet have
spaces in labels because they get used as commands/lookup keys
inconsistently, and we cannot rename FT8 to FTx because it affects
too much, because of labels being misused. But, FT8 settings also
affect FT4 now, so that's a misleading UI.

sbitxd is going to have to break compatibility with legacy .ini files
and commands at some point.