Kindof fix zbitx front panel display: remove tenth-seconds
But the front-panel code is expecting to find 00, 15, 30 or 45 in
specific character positions, and then turning that into a slot number.
It would be much better to have the list be like the sbitx console:
display what it's told to display, and let the main software do the
reformatting to fit the zbitx max width; or else if we could send
messages and semantics as it's done in sbitx already. At least for now,
it's working again, but it thinks every message is in slot 4.

The .uf2 file is from
https://github.com/ec1oud/zbitxfrontpanel/commit/56e4542686de6415b26f3c1d9f38713e0d0abe92